+++ eth_defi/tokenised_fund/midas/__init__.py
@@ -1,0 +1,1 @@
+"""Midas tokenised-fund backfill registration."""